Generate Makefile.am.libasn1codec compatible with non-recursive Makefile.am
The fact that the existing code creates files with names like:
  Makefile.am.libasncodec
  Makefile.am.example

leads me to believe that there is an attempt to create automake compatible
Makefile.am files. However, there are some things that are wrong with the
example generated files. For example, ASN_MODULES_SOURCES variable can't
be used because all *_SOURCES variables are expected to be either a
bin_PROGRAM or lib_LTLIBRARIES, and of course, there is neither:
  bin_PROGRAMS += ASN_MODULES
or
  lib_LTLIBRARIES += ASN_MODULES

This patch does the following:
1) correct the invalid variable names for use with autotools
2) cleanly separate the library generation from an example as1convert prog
3) Add a new option -D <destdir> that will write the generated files to the
   specified directory, and the generated Makefile.am.* will be compatible
   with a non-recursive Makefile.am schema
4) Add a new option -gen-autotools which will create (if they don't exist) a minimal
   example configure.ac and Makefile.am