Remove static qualifier from var in SET_OF__encode_sorted
The function `SET_OF__encode_sorted` stored the working element buffer
in a static variable. If called recursively (that is, when performing
encoding of a SET containing another SET) the outermost element buffer
is therefore clobbered by the innermost call.

This causes a SEGV on continuation of the outermost call.

Bug found during testing of [https://github.com/surevine/spiffing Spiffing],
fix verified with the same test suite.